+++ python/GafferScene/__init__.py
@@ -43,5 +43,6 @@ __import__( "GafferImage" )
 from ._GafferScene import *
 
 from .ShaderBall import ShaderBall
+from .PassWedge import PassWedge
 
 __import__( "IECore" ).loadConfig( "GAFFER_STARTUP_PATHS", subdirectory = "GafferScene" )