+++ python/GafferSceneTest/__init__.py
@@ -175,6 +175,7 @@ from .MeshNormalsTest import MeshNormalsTest
 from .ImageScatterTest import ImageScatterTest
 from .PassesTest import PassesTest
 from .DeletePassesTest import DeletePassesTest
+from .PassWedgeTest import PassWedgeTest
 
 from .IECoreScenePreviewTest import *
 from .IECoreGLPreviewTest import *