+++ python/GafferSceneUI/__init__.py
@@ -189,6 +189,7 @@ from . import LightToolUI
 from . import ImageScatterUI
 from . import PassesUI
 from . import DeletePassesUI
+from . import PassWedgeUI
 
 # then all the PathPreviewWidgets. note that the order
 # of import controls the order of display.